huge refactor: switched to using threading.Event member 'session_up' in place of all tests of session_name is not None, and really rethought and cleaned up the do_io() method's IO event loop.  It is sane at last.
With this commit, the test programs all work, and I don't know of any bugs in the threaded or non-threaded
implementations (other than some thread race conditions in disconnect() that I'm about to tackle).

The code looks much cleaner overall in dealing with the connection state.  Still lots of commented out
cruft though.