major cleanup of the ways a connection is closed.  handle_close and handle_error() now use self._drop() better, and _drop() itself is far better. also added new self._is_context_callback() to make it easier for front-facing methods to tell if they've been invoked from within a handler, and NOT dip back into the event loop recursively. This makes the API re-entrant now.
The do_restart() method is starting to work reentrantly, only for threaded so far, but soon
for non-threaded.

removed threaded usage of self.run() and switched to self.poll().  mostly because run() doesn't
add much and will be front facing only i think

cleaned up initialization of the io_thread
doing lots of work in start_connect() still to make it easier to reconnect a lost session from 
a non-threaded use.