First major attempt to convert the code to work in Python3.
I'm sure this isn't 100% done, but the big things are fixed: print() instead of print
and in many places, .decode() is called now to convert bytes to strings types.
I'm reasonably sure there are latent bytes/string conversion problems still in the code,
as this was just a quick stab to get the example programs working in both python2.4 and python3.1.2

Good news: the example code now works in python 2.4 and python 3.1!